+++ source/loader/ze_loader.cpp
@@ -68,12 +68,34 @@ namespace loader
         return a.driverType < b.driverType;
     }
 
-    bool context_t::driverSorting(driver_vector_t *drivers, ze_init_driver_type_desc_t* desc) {
+    bool context_t::driverSorting(driver_vector_t *drivers, ze_init_driver_type_desc_t* desc, bool sysmanOnly) {
         ze_init_driver_type_desc_t permissiveDesc = {};
         permissiveDesc.stype = ZE_STRUCTURE_TYPE_INIT_DRIVER_TYPE_DESC;
         permissiveDesc.pNext = nullptr;
         permissiveDesc.flags = UINT32_MAX;
         for (auto &driver : *drivers) {
+            if (sysmanOnly) {
+                for (auto &coreDriver : this->zeDrivers) {
+                    if (coreDriver.name == driver.name) {
+                        if (!driver.dditable.ze.Global.pfnInitDrivers) {
+                            driver.dditable.ze.Global.pfnInitDrivers = coreDriver.dditable.ze.Global.pfnInitDrivers;
+                        }
+                        if (!driver.dditable.ze.Driver.pfnGet) {
+                            driver.dditable.ze.Driver.pfnGet = coreDriver.dditable.ze.Driver.pfnGet;
+                        }
+                        if (!driver.dditable.ze.Driver.pfnGetProperties) {
+                            driver.dditable.ze.Driver.pfnGetProperties = coreDriver.dditable.ze.Driver.pfnGetProperties;
+                        }
+                        if (!driver.dditable.ze.Device.pfnGet) {
+                            driver.dditable.ze.Device.pfnGet = coreDriver.dditable.ze.Device.pfnGet;
+                        }
+                        if (!driver.dditable.ze.Device.pfnGetProperties) {
+                            driver.dditable.ze.Device.pfnGetProperties = coreDriver.dditable.ze.Device.pfnGetProperties;
+                        }
+                        break;
+                    }
+                }
+            }
             uint32_t pCount = 0;
             std::vector<ze_driver_handle_t> driverHandles;
             ze_result_t res = ZE_RESULT_SUCCESS;
@@ -356,8 +378,8 @@ namespace loader
             // Use the previously init ddi table pointer to zesInit to allow for intercept of the zesInit calls
             ze_result_t res = sysmanGlobalInitStored->pfnInit(flags);
             // Verify that this driver successfully init in the call above.
-            if (driver.initStatus != ZE_RESULT_SUCCESS) {
-                res = driver.initStatus;
+            if (driver.initSysManStatus != ZE_RESULT_SUCCESS) {
+                res = driver.initSysManStatus;
             }
             if (debugTraceEnabled) {
                 std::string message = "init driver " + driver.name + " zesInit(" + loader::to_string(flags) + ") returning ";